+++ tse_analytics/modules/phenomaster/models/dataset_tree_item.py
@@ -12,10 +12,6 @@ class DatasetTreeItem(TreeItem):
         self.dataset = dataset
         self.meta = dataset.meta
 
-    def rename_dataset(self, name: str) -> None:
-        self.dataset.name = name
-        self.name = name
-
     @property
     def icon(self):
         return QIcon(":/icons/icons8-sheets-16.png")