+++ tse_analytics/views/datasets/adjust_dataset_dialog_ui.py
@@ -15,20 +15,49 @@ from PySide6.QtGui import (QBrush, QColor, QConicalGradient, QCursor,
     QFont, QFontDatabase, QGradient, QIcon,
     QImage, QKeySequence, QLinearGradient, QPainter,
     QPalette, QPixmap, QRadialGradient, QTransform)
-from PySide6.QtWidgets import (QAbstractButton, QApplication, QDateTimeEdit, QDialog,
-    QDialogButtonBox, QGroupBox, QHBoxLayout, QLabel,
-    QRadioButton, QSizePolicy, QSpinBox, QTimeEdit,
-    QVBoxLayout, QWidget)
+from PySide6.QtWidgets import (QAbstractButton, QAbstractItemView, QApplication, QDateTimeEdit,
+    QDialog, QDialogButtonBox, QGroupBox, QHBoxLayout,
+    QHeaderView, QLabel, QLineEdit, QRadioButton,
+    QSizePolicy, QSpinBox, QTableWidget, QTableWidgetItem,
+    QTimeEdit, QVBoxLayout, QWidget)
 import resources_rc
 
 class Ui_AdjustDatasetDialog(object):
     def setupUi(self, AdjustDatasetDialog):
         if not AdjustDatasetDialog.objectName():
             AdjustDatasetDialog.setObjectName(u"AdjustDatasetDialog")
-        AdjustDatasetDialog.resize(362, 264)
+        AdjustDatasetDialog.resize(933, 429)
         self.verticalLayout = QVBoxLayout(AdjustDatasetDialog)
         self.verticalLayout.setObjectName(u"verticalLayout")
-        self.groupBoxResampling = QGroupBox(AdjustDatasetDialog)
+        self.widgetMain = QWidget(AdjustDatasetDialog)
+        self.widgetMain.setObjectName(u"widgetMain")
+        self.horizontalLayout_4 = QHBoxLayout(self.widgetMain)
+        self.horizontalLayout_4.setObjectName(u"horizontalLayout_4")
+        self.widgetLeft = QWidget(self.widgetMain)
+        self.widgetLeft.setObjectName(u"widgetLeft")
+        sizePolicy = QSizePolicy(QSizePolicy.Policy.Fixed, QSizePolicy.Policy.Preferred)
+        sizePolicy.setHorizontalStretch(0)
+        sizePolicy.setVerticalStretch(0)
+        sizePolicy.setHeightForWidth(self.widgetLeft.sizePolicy().hasHeightForWidth())
+        self.widgetLeft.setSizePolicy(sizePolicy)
+        self.verticalLayout_4 = QVBoxLayout(self.widgetLeft)
+        self.verticalLayout_4.setObjectName(u"verticalLayout_4")
+        self.verticalLayout_4.setContentsMargins(0, 0, 0, 0)
+        self.groupBoxRename = QGroupBox(self.widgetLeft)
+        self.groupBoxRename.setObjectName(u"groupBoxRename")
+        self.groupBoxRename.setCheckable(True)
+        self.groupBoxRename.setChecked(False)
+        self.verticalLayout_3 = QVBoxLayout(self.groupBoxRename)
+        self.verticalLayout_3.setObjectName(u"verticalLayout_3")
+        self.lineEditName = QLineEdit(self.groupBoxRename)
+        self.lineEditName.setObjectName(u"lineEditName")
+
+        self.verticalLayout_3.addWidget(self.lineEditName)
+
+
+        self.verticalLayout_4.addWidget(self.groupBoxRename)
+
+        self.groupBoxResampling = QGroupBox(self.widgetLeft)
         self.groupBoxResampling.setObjectName(u"groupBoxResampling")
         self.groupBoxResampling.setCheckable(True)
         self.groupBoxResampling.setChecked(False)
@@ -40,9 +69,9 @@ class Ui_AdjustDatasetDialog(object):
         self.verticalLayout_2.addWidget(self.timeEditResamplingInterval)
 
 
-        self.verticalLayout.addWidget(self.groupBoxResampling)
+        self.verticalLayout_4.addWidget(self.groupBoxResampling)
 
-        self.groupBoxTimeShift = QGroupBox(AdjustDatasetDialog)
+        self.groupBoxTimeShift = QGroupBox(self.widgetLeft)
         self.groupBoxTimeShift.setObjectName(u"groupBoxTimeShift")
         self.groupBoxTimeShift.setCheckable(True)
         self.groupBoxTimeShift.setChecked(False)
@@ -87,9 +116,9 @@ class Ui_AdjustDatasetDialog(object):
         self.horizontalLayout.addWidget(self.radioButtonTimeShiftPlus)
 
 
-        self.verticalLayout.addWidget(self.groupBoxTimeShift)
+        self.verticalLayout_4.addWidget(self.groupBoxTimeShift)
 
-        self.groupBoxTrimTime = QGroupBox(AdjustDatasetDialog)
+        self.groupBoxTrimTime = QGroupBox(self.widgetLeft)
         self.groupBoxTrimTime.setObjectName(u"groupBoxTrimTime")
         self.groupBoxTrimTime.setCheckable(True)
         self.groupBoxTrimTime.setChecked(False)
@@ -100,23 +129,84 @@ class Ui_AdjustDatasetDialog(object):
 
         self.horizontalLayout_2.addWidget(self.labelTrimTimeStart)
 
-        self.dateTimeEditStart = QDateTimeEdit(self.groupBoxTrimTime)
-        self.dateTimeEditStart.setObjectName(u"dateTimeEditStart")
+        self.dateTimeEditTrimStart = QDateTimeEdit(self.groupBoxTrimTime)
+        self.dateTimeEditTrimStart.setObjectName(u"dateTimeEditTrimStart")
 
-        self.horizontalLayout_2.addWidget(self.dateTimeEditStart)
+        self.horizontalLayout_2.addWidget(self.dateTimeEditTrimStart)
 
         self.labelTrimTimeEnd = QLabel(self.groupBoxTrimTime)
         self.labelTrimTimeEnd.setObjectName(u"labelTrimTimeEnd")
 
         self.horizontalLayout_2.addWidget(self.labelTrimTimeEnd)
 
-        self.dateTimeEditEnd = QDateTimeEdit(self.groupBoxTrimTime)
-        self.dateTimeEditEnd.setObjectName(u"dateTimeEditEnd")
+        self.dateTimeEditTrimEnd = QDateTimeEdit(self.groupBoxTrimTime)
+        self.dateTimeEditTrimEnd.setObjectName(u"dateTimeEditTrimEnd")
 
-        self.horizontalLayout_2.addWidget(self.dateTimeEditEnd)
+        self.horizontalLayout_2.addWidget(self.dateTimeEditTrimEnd)
 
 
-        self.verticalLayout.addWidget(self.groupBoxTrimTime)
+        self.verticalLayout_4.addWidget(self.groupBoxTrimTime)
+
+        self.groupBoxExcludeTime = QGroupBox(self.widgetLeft)
+        self.groupBoxExcludeTime.setObjectName(u"groupBoxExcludeTime")
+        self.groupBoxExcludeTime.setCheckable(True)
+        self.groupBoxExcludeTime.setChecked(False)
+        self.horizontalLayout_3 = QHBoxLayout(self.groupBoxExcludeTime)
+        self.horizontalLayout_3.setObjectName(u"horizontalLayout_3")
+        self.labelExcludeTimeStart = QLabel(self.groupBoxExcludeTime)
+        self.labelExcludeTimeStart.setObjectName(u"labelExcludeTimeStart")
+
+        self.horizontalLayout_3.addWidget(self.labelExcludeTimeStart)
+
+        self.dateTimeEditExcludeStart = QDateTimeEdit(self.groupBoxExcludeTime)
+        self.dateTimeEditExcludeStart.setObjectName(u"dateTimeEditExcludeStart")
+
+        self.horizontalLayout_3.addWidget(self.dateTimeEditExcludeStart)
+
+        self.labelExcludeTimeEnd = QLabel(self.groupBoxExcludeTime)
+        self.labelExcludeTimeEnd.setObjectName(u"labelExcludeTimeEnd")
+
+        self.horizontalLayout_3.addWidget(self.labelExcludeTimeEnd)
+
+        self.dateTimeEditExcludeEnd = QDateTimeEdit(self.groupBoxExcludeTime)
+        self.dateTimeEditExcludeEnd.setObjectName(u"dateTimeEditExcludeEnd")
+
+        self.horizontalLayout_3.addWidget(self.dateTimeEditExcludeEnd)
+
+
+        self.verticalLayout_4.addWidget(self.groupBoxExcludeTime)
+
+
+        self.horizontalLayout_4.addWidget(self.widgetLeft)
+
+        self.groupBoxExcludeAnimals = QGroupBox(self.widgetMain)
+        self.groupBoxExcludeAnimals.setObjectName(u"groupBoxExcludeAnimals")
+        self.groupBoxExcludeAnimals.setCheckable(True)
+        self.groupBoxExcludeAnimals.setChecked(False)
+        self.verticalLayout_5 = QVBoxLayout(self.groupBoxExcludeAnimals)
+        self.verticalLayout_5.setObjectName(u"verticalLayout_5")
+        self.tableWidgetAnimals = QTableWidget(self.groupBoxExcludeAnimals)
+        if (self.tableWidgetAnimals.columnCount() < 6):
+            self.tableWidgetAnimals.setColumnCount(6)
+        self.tableWidgetAnimals.setObjectName(u"tableWidgetAnimals")
+        self.tableWidgetAnimals.setEditTriggers(QAbstractItemView.NoEditTriggers)
+        self.tableWidgetAnimals.setSelectionMode(QAbstractItemView.MultiSelection)
+        self.tableWidgetAnimals.setSelectionBehavior(QAbstractItemView.SelectRows)
+        self.tableWidgetAnimals.setVerticalScrollMode(QAbstractItemView.ScrollPerPixel)
+        self.tableWidgetAnimals.setHorizontalScrollMode(QAbstractItemView.ScrollPerPixel)
+        self.tableWidgetAnimals.setSortingEnabled(True)
+        self.tableWidgetAnimals.setColumnCount(6)
+        self.tableWidgetAnimals.verticalHeader().setVisible(False)
+        self.tableWidgetAnimals.verticalHeader().setMinimumSectionSize(20)
+        self.tableWidgetAnimals.verticalHeader().setDefaultSectionSize(20)
+
+        self.verticalLayout_5.addWidget(self.tableWidgetAnimals)
+
+
+        self.horizontalLayout_4.addWidget(self.groupBoxExcludeAnimals)
+
+
+        self.verticalLayout.addWidget(self.widgetMain)
 
         self.buttonBox = QDialogButtonBox(AdjustDatasetDialog)
         self.buttonBox.setObjectName(u"buttonBox")
@@ -135,6 +225,7 @@ class Ui_AdjustDatasetDialog(object):
 
     def retranslateUi(self, AdjustDatasetDialog):
         AdjustDatasetDialog.setWindowTitle(QCoreApplication.translate("AdjustDatasetDialog", u"Adjust Dataset", None))
+        self.groupBoxRename.setTitle(QCoreApplication.translate("AdjustDatasetDialog", u"Rename", None))
         self.groupBoxResampling.setTitle(QCoreApplication.translate("AdjustDatasetDialog", u"Apply resampling", None))
         self.timeEditResamplingInterval.setDisplayFormat(QCoreApplication.translate("AdjustDatasetDialog", u"HH:mm:ss", None))
         self.groupBoxTimeShift.setTitle(QCoreApplication.translate("AdjustDatasetDialog", u"Apply time shift", None))
@@ -143,8 +234,14 @@ class Ui_AdjustDatasetDialog(object):
         self.timeEditTimeShift.setDisplayFormat(QCoreApplication.translate("AdjustDatasetDialog", u"HH:mm:ss", None))
         self.groupBoxTrimTime.setTitle(QCoreApplication.translate("AdjustDatasetDialog", u"Trim time", None))
         self.labelTrimTimeStart.setText(QCoreApplication.translate("AdjustDatasetDialog", u"Start:", None))
-        self.dateTimeEditStart.setDisplayFormat(QCoreApplication.translate("AdjustDatasetDialog", u"yyyy.MM.dd HH:mm:ss", None))
+        self.dateTimeEditTrimStart.setDisplayFormat(QCoreApplication.translate("AdjustDatasetDialog", u"yyyy-MM-dd HH:mm:ss", None))
         self.labelTrimTimeEnd.setText(QCoreApplication.translate("AdjustDatasetDialog", u"End:", None))
-        self.dateTimeEditEnd.setDisplayFormat(QCoreApplication.translate("AdjustDatasetDialog", u"yyyy.MM.dd HH:mm:ss", None))
+        self.dateTimeEditTrimEnd.setDisplayFormat(QCoreApplication.translate("AdjustDatasetDialog", u"yyyy-MM-dd HH:mm:ss", None))
+        self.groupBoxExcludeTime.setTitle(QCoreApplication.translate("AdjustDatasetDialog", u"Exclude time", None))
+        self.labelExcludeTimeStart.setText(QCoreApplication.translate("AdjustDatasetDialog", u"Start:", None))
+        self.dateTimeEditExcludeStart.setDisplayFormat(QCoreApplication.translate("AdjustDatasetDialog", u"yyyy-MM-dd HH:mm:ss", None))
+        self.labelExcludeTimeEnd.setText(QCoreApplication.translate("AdjustDatasetDialog", u"End:", None))
+        self.dateTimeEditExcludeEnd.setDisplayFormat(QCoreApplication.translate("AdjustDatasetDialog", u"yyyy-MM-dd HH:mm:ss", None))
+        self.groupBoxExcludeAnimals.setTitle(QCoreApplication.translate("AdjustDatasetDialog", u"Exclude animals", None))
     # retranslateUi